+++ packages/components-native/src/Form/Form.tsx
@@ -27,6 +27,7 @@ import { FormSaveButton } from "./components/FormSaveButton";
 import { useSaveButtonPosition } from "./hooks/useSaveButtonPosition";
 import { FormCache } from "./components/FormCache/FormCache";
 import { useAtlantisFormContext } from "./context/AtlantisFormContext";
+import { IOSKeyboardAwareScrollViewSpacer } from "./components/IOSKeyboardAwareScrollViewSpacer/IOSKeyboardAwareScrollViewSpacer";
 import { InputAccessoriesProvider } from "../InputText";
 import { tokens } from "../utils/design";
 import { ErrorMessageProvider } from "../ErrorMessageWrapper";
@@ -65,6 +66,7 @@ function InternalForm<T extends FieldValues, S>({
   secondaryActions,
   saveButtonOffset,
   showStickySaveButton = false,
+  disableKeyboardAwareScroll = false,
   renderFooter,
   UNSAFE_allowDiscardLocalCacheWhenOffline,
 }: InternalFormProps<T, S>) {
@@ -165,7 +167,7 @@ function InternalForm<T extends FieldValues, S>({
         >
           <KeyboardAwareScrollView
             enableResetScrollToCoords={false}
-            enableAutomaticScroll={true}
+            enableAutomaticScroll={!disableKeyboardAwareScroll}
             enableOnAndroid={edgeToEdgeEnabled}
             keyboardOpeningTime={
               Platform.OS === "ios" ? tokens["timing-slowest"] : 0
@@ -214,6 +216,9 @@ function InternalForm<T extends FieldValues, S>({
                           onOpenBottomSheet={() => setIsBottomSheetOpen(true)}
                           onCloseBottomSheet={() => setIsBottomSheetOpen(false)}
                         />
+                      )}
+                      {disableKeyboardAwareScroll && (
+                        <IOSKeyboardAwareScrollViewSpacer />
                       )}
                     </View>
                   )}